feat: milestone 7 — the launch executor
Profiles now run. A "Let's race" button on each game opens a live
preflight checklist driven by the scheduler, with steps executing in
parallel where the graph allows.

Nothing outlives a crash
Utilities started for a session go into a Job Object with
KILL_ON_JOB_CLOSE. If Team Principal exits — cleanly, by crash, or killed
from Task Manager — Windows terminates everything in it. Without that, a
crash mid-session leaves SimHub and a vJoy feeder running and the next
launch finds them up in an unknown state, which is worse than the crash.

The game itself is deliberately not in the job. Killing a sim because the
launcher died would lose a race.

Honesty is mechanical, not editorial
EnsureProcess reports AlreadyRunning when a utility was already up, and
the row's text is derived from that rather than written by hand. A step
whose action does not exist yet — an adapter, a display change — fails
saying which milestone it arrives in, rather than silently passing.

The checklist is a pure view over the event stream and never sequences
anything. That is what will let a self-healing check turn green on its
own by the same path its first result took.

Each step is visible for at least a moment, because a row that flashes
past is unreadable on a go/no-go screen — but the padding is per step and
never delays the run, so independent steps stay concurrent and the total
is unchanged. Cancelling tears down what was started; leaving half a
session running because someone changed their mind is how a launcher
earns distrust.

Protocol launches are restricted to Steam, Epic and Store schemes. A
profile is user data, and handing an arbitrary string to ShellExecute
would make it a way to run anything at all.

Status on every row is glyph, colour and word. This is a go/no-go screen
and colourblind users exist.

Also: StepView.elapsedMs is f64 rather than u64, because a u64 crosses
into JavaScript as bigint and cannot be divided by a number without a
cast at every call site. Same reasoning as the desktop areas earlier.

Profile editing — which utilities to start, which peripherals are
required, which display layout — lands with the full Let's race flow in
milestone 8. The page says so.

+++ crates/tp-model/src/graph.rs
@@ -318,7 +318,10 @@ pub struct StepView {
     pub severity: Severity,
     pub status: StepStatus,
     pub detail: String,
-    pub elapsed_ms: Option<u64>,
+    /// f64 rather than u64: a u64 crosses into JavaScript as `bigint`, which
+    /// cannot be divided by a number without a cast at every call site. Elapsed
+    /// milliseconds never approach the precision limit.
+    pub elapsed_ms: Option<f64>,
     /// What the app actually did, so the row cannot claim credit it has not
     /// earned — "Already running" is derived from this, never written by hand.
     pub action_taken: crate::ActionTaken,
